extract to a separate class
